Added Bullet class

+++ src/Renderer/VertexBufferLayout.cpp
@@ -11,7 +11,7 @@ namespace RenderEngine {
 	}
 	void VertexBufferLayout::addElementLayoutFloat(GLint count, const bool normalized)
 	{
-		m_layoutElements.push_back({count, GL_FLOAT, normalized, count * sizeof(GLfloat)});
+		m_layoutElements.push_back({ count, GL_FLOAT, normalized, count * static_cast<unsigned int>(sizeof(GLfloat)) });
 		m_stride += m_layoutElements.back().size;
 	}
 }
